Allows logging all messages without limitations.
This can be useful for instance in cases where the user wants to export the list of all failed imported entities.
For now, the parameter is collected for all file import jobs as they seem to be the actual use case, where knowing all failures may be relevant to some users.
By default, the logic remains unchanged. For jobs with the parameter, the default is set to enforce the limiting and for jobs without the parameter, the limit is also enforced.

Fixes: SE-12352