BREAKING: Cleanup of the inner workings of the EntityParameter
This resolves some Sonarlint issues. We made as many methods protected as possible.
Also, we reference an Autocompleter by its type in favor of a name to ensure proper
type-safety. Also a method was de-deprecated as it is still possible (and might even
be feasible) to provide a fully custom autocomplete URL.

These changes are breaking in their nature but should have minimal to no impact
to downstream code.