Major package cleanup for all chart and metric related classes.
We killed (only left deprecated classes) in the charts package. Also the DataExplorer was put
in a top-level package in analytics. The only survivor of this massacre was the Dataset class
which has been moved as well. Additionally the "dashboards" package was renamed to
"metrics" as we killed the whole "charts" facility here (as it has been replaced by the
DataExplorer anyway".

Next to tons of import updates and some JavaDoc fixes, this commit contains no semantic
changes its deletes and renames all the way down...