Changes the way global "batches" are executed.
For global metric tasks, we used abstract classes like
MongoEntity or SQLEntity and created a "fake" batch
which was empty otherwise. Additionally, in some cases -
namely 1) in case of an exception and 2) in case of
no entities present at all - no batch was generated at all.

However, having multiple levels we actually rely on at
least one batch being present, so that we can add the
LAST_BATCH marker and thus scheduler the next
level. So, for global runs, we now no longer use an
completely empty batch description, but one with a
special flag being set. Additionally, in the cases named
above, we emit a completely empty batch - which is used
to forward the LAST_BATCH marker but is then safely
discarded.