Returns null for empty StringParameter
This basically restores the previous behavior. If no default was returned, and the parameter is empty (either not in context or provided blank) we want to return a null string. Reason for that is that the output is often used in DB queries where null is expected instead of a empty string.

Fixes: SIRI-713