Moves a special handler into the proper framework layer.
Instead of making the layer3 VirtualFileSystemController aware of Blobs (layer2) and their special
abilities, we moved the proper handling in the L3Uplink which forms the interface between
level2 and level3. This way, we can also get rid of a sort of redundant method (Blob#deliver) as
we now always route all traffic through the BlobDispatcher.