Skips large file delivery if only a range is requested.
If a file or object is requested which is known to be very large,
we used a special delivery method in a separate thread pool
using blocking IO. We did this, as in async mode, our internal
buffers might overflow if a slow client cannot catch-up with
our fast S3 backend when tunneling data.

However, if only a range is requested, the amount of bytes
to shovel is probably limited and also we did not support
range requests at all for the blocking IO. This meant that
we probably transfer way too many bytes which also seems
to kill Safari when requesting an MP4 video file.