Ensures uplink is returned to object pool when exception occurs
Previously an error in completePendingCommand (e.g. when the connector thew an error because the response was not readable or had a bad status code) we would not safe close the connector (after the changes of commit 35df532fddd8d7be0be462205bba52e6bb0e5cc9).
This meant that the uplink would not be reported as idle to the object pool and would never get cleaned up after the idle timeout is reached, instead it was kept open.
Finally this meant that the internal operation of the uplink would hit its operation timeout and a hanging operation is reported.

Fixes: OX-9618