Fixes moving of files via SFTP
PR #1554 implemented renameFile to fix renaming of files via SFTP. Unfortunately, that change broke moving them, as both are the same operation in POSIX. The default implementation uses Files#move, which triggers BridgeFileSystemProvider#move, so it makes sense to handle both cases there.

Some clients (mainly CLI) can move and rename at the same time, but implementing this increases the complexity quite a bit because our framework views both as separate operations. We keep it simple and only support one at a time. For most clients and use-cases this should be fine.

Fixes: SIRI-775