Temporarily reduce the length of logged incidences, mails, and messages
ES < 8.7.0 has a bug were it wrongfully raises an error for  fields longer than 32766 that are not indexed in lucene (see https://github.com/elastic/elasticsearch/pull/93299).

This PR introduces a temporary limit on the length of some commonly long fields of logging entities and lowers the configured default length of LoggedMessage. We use 30k instead of 32766 as ES checks the byte length and not the character count. 30k should be good enough as the number of non-ASCII characters should be quite low.

Fixes: SIRI-786