Permits to attempt the deletion of read-only files
Setting the flag before will not alter the state of a VirtualFile, as it would need to be refreshed (re-wrapped).

Since the deletion block is a logical check on a flag, we ignore it and allow the deleteHandler to try perform the deletion, which might fail or not depending on the underlying storage.

Fixes: SIRI-805