Removes duplicate access path resolves
The target object is already resolved in Property#getValue. Property#getValue then calls Property#getValueFromField. BlobRefProperty#getValueFromField is overridden to return the blob key instead of the BlobHardRef object via BlobRefProperty#getRef. There, we tried to resolve the target object again, which fails when used inside composites. This removes the duplicate resolve. As BlobRefProperty#getRef is used in other places which don't have the target object yet, we moved the resolve there.

Fixes: SIRI-829