Extracts the parsing of Blob URIs
The previous code became unreadable and hard to maintain. Also, we need the same functionality for the next commit. This commit extracts the parsing into a utility class.

We intentionally don't define constants for every URI part, as they have different meanings in the different URIs.

Drive-buy: This fixes marking blobs as touched when using physical URIs (physicalDelivery() expected the blob key as 4th parameter).

Fixes: SIRI-825