Changes the order in which passwords and API-Tokens are checked
We prefer checking the password, as we're trying to migrate away from
API tokens. Therefore we monitor the logs for "AuditLog.apiTokenLogin"
and place these API tokens as properly encrypted and protected password.

Once a system no longer receives these messages, we can disable
the support for api tokens entirely.