Allows to attach kba link to a process log message
This new log handler allows to specify a KBA ID for a ProcessLog message that is rendered as a link below the message.
The link is only rendered when the user can access the article and has the title of the article as link text (like the KB message expander).

Fixes: OX-10350