Allows to easily copy the deeplink of a kba section
This renders a button in the header of a section when an anchor is given as parameter.
When the button is clicked the signed URL (valid for 30 days) is expanded by the anchor, which opens the article directly at the given section.

This works for sections with and without a heading.
When no anchor is given no button is rendered.