Introduces the Blob.duplicate() method
Used to clone a blob. This method delivers a BlobDuplicator class which collects parameters for the copy, such as variants to clone with or a new target storage space.

The new blob created is temporary and must be handled after creation (for example, attaching it to an actual BlobHardRef), otherwise it will be deleted from the system after a while.

Fixes: SIRI-851