Generally shows the "edit" action for users ✏️
Showing the action may be redundant in certain settings, as it could have the same effect as clicking on the entire card. However, a user focusing on the menu may miss the click action entirely. Therefore, we show it in the menu as well. (This is a UI pattern we use in OX as well.)

SIRI-852