Also shows a "select" action in "select" mode 🚂
Similar to commit 435defff311487d316135b3d182d57ee4efdc69a, the main action is now also shown as first action in the menu. Again, the idea is to make things more obvious to a user focusing on the menu only. (This is a UI pattern we use in OX as well.)

SIRI-852