POSSIBLY BREAKING: Do not pre-fill multi select fields in jobs.
The optional parameter should be (as the old comment suggested) always be set to true, as otherwise the field will be filled with the first possible value, which is not the expected behaviour.

This change is breaking if 'paramMultiSelect.html.pasta' is called from another template, which is not the case in the Sirius framework and unlikely in general.

Fixes: SIRI-873