Only warns about a failing update task when deletion is queued
Normally this class has a mechanism that deletes previous tasks when a new task is created. Regardless this failed for a few files and an update task existed even though a deletion also existed.
This means the update task cant succeed as the file is already removed from the main storage.
To reduce the amount of scary error messages in such cases we check whether a equivalent deletion task exists when an update task fails.
If that is the case we only output a warning (so its still traceable if noticed).
Also we mark the task as failed to prevent the loop from re-trying the update.

Fixes: SIRI-878