Keeps a separate history for the data explorer
Previously the data explorer messed with the tycho history and did not handle reloads and back navigation very well.
Now the data explorer writes his own states and tries to apply the state of the URL when navigating in the history.
This includes showing the overview, adding charts and removing charts.
For this to work some methods had to be expanded by a silent parameter, which skips the URL and history update. These are used whenever the page load or history state handling are calling the methods. Only the UI elements should update the URL and history state. Otherwise reloads, etc. would introduce multiple duplicate entries in the browser history.

Fixes: OX-791