Adds helpers for redirecting calls to safe referers 🚧
We often have the combination of an overview page `/entities` and respective details pages like `/entity/123`. In Tycho, we may want to provide certain actions that need to work from both locations, for instance archiving or unarchiving. The respective routes need to redirect smartly to wherever the call came from.

This commit adds methods that redirects back to the referer, maintaining possible query parameters, after having checked it.