Resolves current tenant directly from `UserContext` 🕵️‍♂️
The previous implementation used `getCurrentUser()` which resolves a database object. However, if we use synthetic users (as we do, for instance, in `runAsAdminOfTenant(…)`), there is no database entry, and this method always returned **null**. A `UserInfo` object on the other hand can always be resolved, and along with it a tenant object.