Switches eternallyValid x reusable
This change aims to clearly split URL types from its validity.

An eternally valid URL means that its hash is always valid, independent if used in a virtual or physical URL.

Reusable sort of achieved the same thing, but forced URLs to be virtual. A virtual URL causes a lookup in the system and sometimes you want to deliver a long living physical url.

Fixes: OX-10300