Improves the return of hash verification
This is a breaking change if someone is using this method directly!

Instead of just returning a simple boolean indicating if the hash passes or not, we return the amount of days passed since today to where the hash successfully validated. This is needed later so we can properly calculate the response cache times.

Fixes: OX-10300