Replace Value with List<String>
This requires a bit more handling of empty or single element lists, but has less magic and should be easier to understand.

Also, this fixes fetchParameter to actually return the list instead of a concatenated string. Only asParameterContext is allowed to return those.

Empty parameters are now stored as null, while maintaining the correct logic for parameter defaults.

Fixes: SE-13229