Ignores ConnectionClosedException
This tries to ignore a rare exception with message "Premature end of Content-Length", which seems to happen VERY SPORADICALLY, when streaming data from the S3 storage to the client, where the client connections sort of die during the transfer.

We could correlate this behavior so far with connection from overseas, where a client disconnect was always on-pair with this error message.

Fixes: SIRI-923