Correctly initialises tab bar when in dropdown/multiline mode 💦
The component makes assumptions about the inner workings of Bootstrap that are no longer met by version 5. It turns out that changing the tabs via the drop-down menu (as happens if there are more than 4 languages) does not work initially, due to a conflicting inner state between ourselves and bootstrap. For now, this can be solved by simulating a click on the entry for the selected language. This will bring the states into synch.

SIRI-925