Avoids cursor exception processing analytic computers
By using streamBlockwise instead of interateAll.

Although batches are emitted to only 250 entities by default (can be overwritten), for each entity, all relevant computers are called, causing the cursor to remain open for quite a long time.

We've seen a MongoCursorNotFoundException at a productive MongoDB instance.

Note that mongo queries shall not provide sorting, which is incompatible with streamBlockwise