Correctly prints the original URL 🚑
In order to route back to the originally intended site after a SAML login, the login site stores the original URL in the local storage. This is done by passing the URL to the Pasta template and rendering the string as part of a respective JavaScript snippet. If a URL contains characters with significance in HTML such as `&`, though, the Pasta string renderer converts these to HTML entities unless everything is placed within `<i:raw>`.

SIRI-937