Utilizes the CustomEvents framework in ImportJobs
If there is only one script present, it will be ued right away. Only if there are several present for the current tenant, the user can pick which
should be used. This way, a script will be automatically used everywhere if only one exists. If the script is optional, a second and empty
script can be provided.