Permits to disable scripting in jobs
enableScriptableEvents must be overwritten so scripting takes effect.

If active, and only one script exists, the parameter is collected hidden and the first entry will be the default one.

This also provides better tracking to which dispatcher was actually used (even a hidden one) since the parameter and its value is logged at job start.

Fixes: OX-9839