Improves the tenant select check
When spying over a user WITHOUT the permission to select tenants and navigating to a URL of the same tenant being spied, we would get a permission failed check.

Note that a logged-in user can navigate to URLs of its own tenant, so we just extend that when spying on a foreign user.