Adds tag parameter to control whether to suggest deprecated values or not.
To keep things clear, the actual logic will be implemented in later commits.

Note, that in the templates, the parameter is named "suggestDeprecatedValues" as we might want an additional parameter later named "showDeprecatedValues" for controlling whether deprecated values should show up in the modal.
Contrary, in the controller, the parameter is named "considerDeprecatedValues" as the route uses both suggest() methods and scan() methods.

Fixes: SIRI-975