BREAKING: change the way saving processTypes.
until now we always stored getClass().getSimpleName() + ".label" = MyJobFactory.label in the processType field on processes. For overridden implementations of e.g. abstract jobs this leads to not make the processes filterable on /ps.
Now we will use the job#getName and store it as the processType. So the filtering url no longer is /ps?processType=BmecatImportJobFactory.label but /ps?processType=import-bmecat
To migrate existing values we implement a temporary job which resolves all available jobs and resolves the correct new type.

- fixes: SIRI-990