Improves labelling of a `<k:link>` 🏷️
First, using the body block (i.e. nested elements) to pass the caption of a link feels more HTML-ish. Second, it is probably not well-defined what the `name` of a link is supposed to be. Therefore, this commit introduces a new attribute `label` that defaults to the rendered content of the element. The old `name` attribute is deprecated and used as fallback only.

OX-11128