Introduces configurable and adjustable password requirements (and generated password settings) as well as adjustable info texts.
(Usage of the new class will be introduced later.)

By default, the settings match the currently shown texts and active checks.

Note, that the PasswordSettings class is implemented as a part and is entity specific allowing for custom implementations (that replace this part) where the settings might be different based for different entities / types of entities.

Also note, that the parameter entity must be of type "Entity" and not "BaseEntity" due to some classes like the abstract UserAccount not directly extending BaseEntity.

Fixes: SE-14149