Always redirect to a physical file URL when it exists
and sets caching appropriately: either we don't cache if the URL says so, or we cache it to a maximum or 1h, as the target location of the redirected URL might change.

Previously, we cached the redirect URL as long as the physical URL, and that is usually very long to infinite.

Fixes: OX-11230