Allows blockwise fetching of event entities.
Events have no ID or other distinct fields and multiple events may be recoreded at the same time, fetching a block must start at the time of the last event of the previous batch (inclusively).

To prevent duplicates and because no fields are distinct by default, callers must provide a duplicate preventer based on the type of events they wish to fetch.

Fixes: SE-14230