Introduces a doughnut chart 🍩
In order to make the chart more easily accessible, a new class is introduced. However, the functionality all resides in the pie chart, as a doughnut chart can degenerate to a pie chart.

The renaming was suggested by @sabieber during review. Writing adapted to BE, as is done in Wikipedia.

OX-9272