BREAKING: introduce method to buildUnderlyingQueryWithFacets
the existent method already does this, but if you already have a page, the filter-facet would be applied doubled.
So we simply duplicate the method but without applying the facets and for usages where the facets are needed, now a change is needed.

Patch-Task: change buildUnderlyingQuery to buildUnderlyingQueryWithFacets where facets needed.

- fixes: SIRI-1054