Allows canceling the task context of scripts.
While this doesn't cancel scripts themselves, code that does check for an active task context (e.g. database query iterations) do check for it.

Note, that the KBA already specifies a way to cancel a script when executing a specific line of code via a second script. This however (up to now) only worked for scripts that created a process as that has its own task context adapter which supports cancelling.

Fixes: SIRI-1053