Considers waitLonger when delivering URLs
We detect the "wish" to wait longer and sets a header in the response, which is then read further down in order to control how many retries we want to do.

Using a header and simply changing the number of retries reduces the number of chained methods that must be changed to propagate the information further down.

Fixes: OX-11845