Process blobs in blocks
By pulling a block of blobs and process them in parallel.

This job is expected to run over a long time on systems with large number of blobs, therefore not using the progress tracker as the count might go off. We simply use single counters to at least know the job is alive

Fixes: SIRI-993