Ads a rudimentary restart ability
If the job is aborted, the first ID of the current block being processed is printed, which can be used as new start point.

Since the parallel executor also monitors if the current task context is active, we cannot guarantee which ID was processed last, so this comes with a small trade off that part of the block will be re-checked upon restart.

Also adds safe measures so lines and counters are no longer written to an aborted job

Fixes: SIRI-993