Logs erroneous parameter exceptions when executing job presets via file system uploads or scheduled job execution.
This most likely happens if a new and required parameter is added to a job and not all systems are properly checked for existing presets or scheduled entries.

Note, that the exception handler re-throwing the exception leads to it being caught in the inner #buildAndVerify method again which invokes the exception handler a second time. This results in the exception being logged twice, which however is better than adding a more complex exception handling to avoid a rare edge case exception simply being logged twice.

Also note, that #startInBackground happens very late. Thus we do not know if we come from a job preset or scheduled entry (the logged stack trace however will indicate that). Also, user errors in the backend UI will be checked way earlier which is good as we don't want to log when a user forgets a parameter.

Fixes: SIRI-148