Changes response status when conversion finally failed
The previous response (500 - Internal Server Error) is thrown internally by response.tunnel() when connection errors happen. This code puts a conversion node in black list for 5 minutes.

Failing to convert a variant does not mean the host is down, so we will return 404 not found as a more appropriate code in this case.

Fixes: SIRI-1136