Checks for missing bucket objects
By looping the selected bucket and querying if a blob or variant exists for it.

We collect up to 10.000 objects to query in a queue and use our parallel executor to perform the actual DB lookups in a number of threads specified by the user (up to 4).

Orphan files are logged in the process as a compressed CSV file.

Fixes: SIRI-992