SIRI-1155 add possibility to filter for permissions in custom-properties
the filtering is helpful for having different properties in different tenants.
As not every tenant should see and use all properties we have to filter then out here.
the manyRows command cannot be filtered afterwards, cause a filtering then causes broken pagination in the modal, so we do not use it anymore here.

- fixes: SIRI-1155