Calculates the checksum of uploaded streams
With input streams, things get quite complicated, since the stream will be passed to tons of methods down the line which will finally perform the upload in the target layer1 storage.

Therefore, we encapsulate the original input stream in a FilterInputStream and act on calls to the read methods, passing the read bytes to the hasher as the stream is consumed, does not matter who does it.

Fixes: SIRI-1158