Completes a refactoring in sirius kernel which modernizes Outcall
This change now tries to cache a generated HttpClient as much/long
as possible. Also the calls also the class has been moved.

On our case, we do not want to cache the generated clients for
the VFS, as we provide a specific CookieManager.

For the cluster sync however, the same client can be re-used.