Adds an "Edit" button to KBAs ✍️
In general, editing KBAs is currently a bit cumbersome as true source code files need to be modified that reside within the code repository. In order to facilitate this, this commit adds an edit button to rendered KBAs that is visible only to system tenant users and that directly opens the respective editor on GitHub. This should make it simpler to at least edit existing articles with reduced effort and without the need of an IDE or much knowledge of Git.

There are, however, known caveats: The code uses some rules to guess the repository from the URL of the Pasta-file. If successful, it assumes that it belongs to scireum and uses a `develop` branch. If these assumptions are wrong, though, the edit link may lead to nowhere. For internal projects, however, this assumption seems to be mostly valid.

MIO-6446