Generate category jump links in open search
This renders a list of jump buttons for each category that is rendered for an search term in the open search. Also re-uses the category icons to make them easier to identify/recognize.
Clicking one of the buttons scrolls the view to the top of the category.

This has two main benefits:
- Better overview of what type of results are available
- quicker navigation to the desired result type